add sonar

+++ redmine-restore.sh
@@ -1,8 +1,6 @@
-@echo off
-docker-compose stop redmine
-docker-compose rm -f redmine
-if "%1" == "" (
+docker-compose rm -sf redmine
+if [ $# == 0 ]
     docker-compose run --rm redmine app:backup:restore
-) else (
+else
     docker-compose run --rm redmine app:backup:restore BACKUP=%1
-)
+fi
